GPGGA added to delimit, logic for signs added to latlong
The if statement for a GPGGA outputs the same as the GPRMC, but pulls
from fields 2 and 4 instead of 3 and 5 since GGA does not have the
status indicator field.
The logic added to latlong allows the function to accept coords in any
hemisphere, but this currently makes no change because of the logic
statements in delimit.